Delay arch_disable_mmu()
On gcc versions after 11.2.0, when using -O2 the compiler inlines pm8xxx_reg_write()
function into the others that use it. However, when it does so, on the
assembly `cmd` ends up unaligned. This is fine if MMU is enabled, but
when executing pmic_reset_configure(), inside scm_elexec_call(), MMU is
disabled. This causes a data abort, which causes lk to try and reboot
into DLOAD mode, only to hit the exact same bug on
pm8x41_clear_pmic_watchdog().

Fix this by keeping MMU enabled until almost just before kernel is about
to boot. This way we also avoid potentially hitting the same bug on
another function in the future.